extract DependencyGraph into its own module
